feat: adjust component position
