+++ packages/client/src/collection-manager/Configuration/ImportXlsx/xlsxFormValueContextProvider.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface xlsxFormValueProps {
+  value?: any;
+  setFormValue?: any;
+}
+
+export const xlsxFormValueContext = React.createContext<xlsxFormValueProps>({});
+export const useFormValueContext = () => React.useContext(xlsxFormValueContext);
