+++ packages/client/src/collection-manager/templates/import.tsx
@@ -3,7 +3,7 @@ import { getConfigurableProperties } from './properties';
 
 export class ImportCollectionTemplate extends CollectionTemplate {
   name = 'import';
-  title = '{{t("Import collection")}}';
+  title = '{{t("Import json collection")}}';
   order = 5;
   color = 'blue';
   divider = true;
