+++ packages/client/src/collection-manager/templates/xlsx.tsx
@@ -2,7 +2,7 @@ import { CollectionTemplate } from '../../data-source/collection-template/Collec
 
 export class ImportXlsxTemplate extends CollectionTemplate {
   name = 'importXlsx';
-  title = '{{t("Import xlsx")}}';
+  title = '{{t("Import xlsx collection")}}';
   order = 6;
   color = 'blue';
   divider = false;
